refactor out `collectCommitLabels` function to separate module
for use outside of changelog-maker cause it's quite handy
